+++ CHANGELOG.md
@@ -16,6 +16,8 @@ This update accompanies the release of the [Telegram Bot API 6.6](https://core.t
    using the method `\TgBotLib\Bot > getMyDescription()` see [getMyDescription](https://core.telegram.org/bots/api#getmydescription) for more information.
  * Added the ability to set different bot short descriptions for different user languages using the method `\TgBotLib\Bot > setMyShortDescription()`
    see [setMyShortDescription](https://core.telegram.org/bots/api#setmyshortdescription) for more information.
+ * Added the ability to get the current bot short description in the given language as the class [BotShortDescription](https://core.telegram.org/bots/api#botshortdescription)
+   using the method `\TgBotLib\Bot > getMyShortDescription()` see [getMyShortDescription](https://core.telegram.org/bots/api#getmyshortdescription) for more information.
 
 ### Changed
  * Removed unused `__destruct()` method from `\TgBotLib\Bot`